+++ react-components/src/components/Sidebar/CountryNotSelected/index.jsx
@@ -1,14 +1,13 @@
-import React, { useState } from 'react'
+import React, { useState, memo } from 'react'
 import PropTypes from 'prop-types'
 import ReactModal from 'react-modal'
 
-import { Modal } from '@src/components/Modal/index'
-import CountryFinderModal from '../../ProductFinder/CountryFinderModal'
+import { Modal } from '@src/components/Modal'
+import CountryFinderModal from '@src/components/ProductFinder/CountryFinderModal'
 import Services from '@src/Services'
 import actions from '@src/actions'
 
-export const CountryNotSelected = ({ isOpen }) => {
-
+export const CountryNotSelected = memo(({ isOpen }) => {
   const [modal, setModal] = useState(isOpen)
   const [modalIsOpen, setIsOpen] = useState(false)
 
@@ -25,17 +24,17 @@ export const CountryNotSelected = ({ isOpen }) => {
     <>
       <ReactModal
         isOpen={modal}
-        className='ReactModal__Content ReactModalCentreScreen'
-        overlayClassName='ReactModal__Overlay ReactModalCentreScreen'
-        contentLabel='Modal'
+        className="ReactModal__Content ReactModalCentreScreen"
+        overlayClassName="ReactModal__Overlay ReactModalCentreScreen"
+        contentLabel="Modal"
       >
         <Modal
-          backUrl='/export-plan/dashboard/'
-          header='Add your target market'
-          content='You will need to choose a target market before you can complete this section'
+          backUrl="/export-plan/dashboard/"
+          header="Add your target market"
+          content="You will need to choose a target market before you can complete this section"
           onClick={openCountryFinder}
-          buttonText='Add a target market'
-          type='1'
+          buttonText="Add a target market"
+          type="1"
         />
       </ReactModal>
       <CountryFinderModal
@@ -45,7 +44,7 @@ export const CountryNotSelected = ({ isOpen }) => {
       />
     </>
   )
-}
+})
 
 CountryNotSelected.propTypes = {
   isOpen: PropTypes.bool.isRequired,